git-svn-id: http://streamboard.de.vu/svn/oscam/branches/sc8in1@6300 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ csctapi/ifd_sc8in1.h
@@ -21,7 +21,7 @@
 int32_t Sc8in1_Init (struct s_reader * reader);
 int32_t Sc8in1_GetStatus (struct s_reader * reader, int32_t * status);
 int32_t Sc8in1_Card_Changed (struct s_reader * reader);
-int32_t Sc8in1_Selectslot(struct s_reader * reader, int32_t slot);
+int32_t Sc8in1_Selectslot(struct s_reader * reader, uint16_t slot);
 int32_t Sc8in1_GetActiveHandle(struct s_reader *reader);
 int32_t Sc8in1_Close(struct s_reader *reader);
 int32_t Sc8in1_InitLocks(struct s_reader * reader);
